Consolidate templated elem instance creation for reuse
